chore(angular): move lib tests from extensions to core
